Solving tsc type errors
