+++ packages/db-common/internal.ts
@@ -710,6 +710,7 @@ export type SelectOptions = {
   description?: string;
   recursive?: boolean; // for File.find()
   client?: DatabaseClient;
+  schema?: any;
 };
 export const orderByIsObject = (
   object: any
@@ -803,7 +804,7 @@ export type StrongRow = { [key: string]: Value };
 export type PrimaryKeyValue = number | string;
 
 //https://stackoverflow.com/a/57390160/19839414
-export type PartialSome<T, K extends keyof T> = Partial<T> & Pick<T, K>
+export type PartialSome<T, K extends keyof T> = Partial<T> & Pick<T, K>;
 
 export const prefixFieldsInWhere = (inputWhere: any, tablePrefix: string) => {
   if (!inputWhere) return {};
