+++ packages/saltcorn-data/models/trigger.ts
@@ -232,7 +232,9 @@ class Trigger implements AbstractTrigger {
       return;
     }
     setTimeout(async () => {
-      const { password, ...user } = userPW as User;
+      // const { password, ...user } = userPW as User;
+
+      const { password, ...user } = (userPW || {}) as User;
       const { getState } = require("../db/state");
       if (!getState) return; // probably in a test
       const findArgs: Where = { when_trigger: eventType };
